ajustes inde y main
