+++ src/lib/index.js
@@ -1,6 +1,20 @@
 // aqui exportaras las funciones que necesites
 
-export const myFunction = () => {
-  // aqui tu codigo
-  console.log('Hola mundo!');
+// export const myFunction = () => {
+// aqui tu codigo
+// console.log('Hola mundo!');
+// };
+
+// Your web app's Firebase configuration
+// For Firebase JS SDK v7.20.0 and later, measurementId is optional
+const firebaseConfig = {
+  apiKey: 'AIzaSyDsV641BeYPAlGnaSM_CPuV5nWGVNOiPZs',
+  authDomain: 'catslover012023.firebaseapp.com',
+  databaseURL: 'https://catslover012023-default-rtdb.firebaseio.com',
+  projectId: 'catslover012023',
+  storageBucket: 'catslover012023.appspot.com',
+  messagingSenderId: '262132831591',
+  appId: '1:262132831591:web:12a097d59110e034244bc2',
+  measurementId: 'G-K4QCS1CZVP',
 };
+export default firebaseConfig;
