+++ src/main.js
@@ -6,30 +6,19 @@ import {
   signInWithPopup,
 } from 'firebase/auth';
 import { initializeApp } from 'firebase/app';
+import firebaseConfig from './lib/index';
 // import { getAnalytics } from 'firebase/analytics';
-import { myFunction } from './lib/index.js';
+// import { myFunction } from './lib/index.js';
 
 // const analytics = getAnalytics(app);
 
 // TODO: Add SDKs for Firebase products that you want to use
 // https://firebase.google.com/docs/web/setup#available-libraries
 
-// Your web app's Firebase configuration
-// For Firebase JS SDK v7.20.0 and later, measurementId is optional
-const firebaseConfig = {
-  apiKey: 'AIzaSyDsV641BeYPAlGnaSM_CPuV5nWGVNOiPZs',
-  authDomain: 'catslover012023.firebaseapp.com',
-  databaseURL: 'https://catslover012023-default-rtdb.firebaseio.com',
-  projectId: 'catslover012023',
-  storageBucket: 'catslover012023.appspot.com',
-  messagingSenderId: '262132831591',
-  appId: '1:262132831591:web:12a097d59110e034244bc2',
-  measurementId: 'G-K4QCS1CZVP',
-};
 const btnGoogle = document.getElementById('btnGoogle');
 // Initialize Firebase
 const app = initializeApp(firebaseConfig);
-console.log(app);
+//console.log(app);
 
 btnGoogle.addEventListener('click', () => {
   const auth = getAuth();
@@ -56,9 +45,12 @@ btnGoogle.addEventListener('click', () => {
 
 const auth = getAuth(app);
 
-myFunction();
+// myFunction();
 
 const btnLogoutGoogle = document.getElementById('btnLogout');
 btnLogoutGoogle.addEventListener('click', () => {
   auth.signOut();
+  console.log('cerraste sesion');
+  console.log(auth);
+  console.log(document.cookie);
 });
